+++ osf_scraper_api/Dockerfile
@@ -1,4 +1,4 @@
-FROM ubuntu:14.04
+FROM osf_scraper_base
 
 # different builds for different environment
 ARG BUILD_ENV
@@ -8,63 +8,27 @@ RUN echo "++ building Dockerfile for env: ${BUILD_ENV}"
 ENV APP_NAME osf_scraper_api
 ENV REPO_URL git@github.com:mhfowler/open-source-feeds.git
 ENV REPO_REMOTE origin
-ENV SRV_DIR /srv
-ENV LOG_DIR /srv/log
-ENV DATA_DIR /srv/data
 ENV PUBLIC_DIR /srv/public
 ENV SRC_DIR /srv/app
 ENV WDIR /srv/app/osf_scraper_api
-ENV ENV_PATH "devops/secret_files/env/${BUILD_ENV}"
-
-# Update OS
-RUN apt-get update
-RUN apt-get -y upgrade
-
-# Install Python
-RUN apt-get install -y python-dev python-pip
-RUN pip install --upgrade pip
-
-# Install other dependencies
-RUN apt-get install -y \
-    libpq-dev \
-    nginx \
-    git \
-    curl
+ENV BASE_DIR .
+ENV DEVOPS_DIR $BASEDIR/osf_scraper_api/devops
+ENV ENV_PATH "${LOCAL_DIR}/devops/secret_files/env/${BUILD_ENV}"
 
 # copy over nginx conf
-COPY devops/build/nginx.conf /etc/nginx/nginx.conf
-COPY devops/build/nginx_site.conf /etc/nginx/sites-available/app.conf
+COPY ${DEVOPS_DIR}/build/nginx.conf /etc/nginx/nginx.conf
+COPY ${DEVOPS_DIR}/build/nginx_site.conf /etc/nginx/sites-available/app.conf
 RUN rm /etc/nginx/sites-enabled/*
 RUN ln -s \
     /etc/nginx/sites-available/app.conf \
     /etc/nginx/sites-enabled/app.conf
 
-# create user groups
-RUN groupadd webgroup
-RUN usermod -G webgroup root
-RUN useradd -G webgroup ubuntu
-RUN useradd -G webgroup nginx-user
-RUN useradd -G webgroup uwsgi-user
-
-# copy over ssh keys
-COPY devops/secret_files/osf_rsa.pub /root/.ssh/id_rsa.pub
-COPY devops/secret_files/osf_rsa /root/.ssh/id_rsa
-
-# ensure directories
-RUN mkdir $LOG_DIR
-RUN mkdir $DATA_DIR
-
 # copy code
-COPY . $SRC_DIR
+COPY $BASEDIR/osf_scraper_api/osf_scraper_api $SRC_DIR/osf_scraper_api
+COPY $BASEDIR/osf/osf $SRC_DIR/osf
 
 # copy hellow_webapp.ini
-COPY devops/build/app.ini "${SRC_DIR}/app.ini"
-
-# Install Supervisord
-RUN apt-get update && apt-get install -y supervisor \
-&& rm -rf /var/lib/apt/lists/*
-# Custom Supervisord config
-COPY devops/templates/supervisord.conf /etc/supervisor/conf.d/supervisord.conf
+COPY ${DEVOPS_DIR}/build/app.ini "${SRC_DIR}/app.ini"
 
 # copy env.json
 RUN echo "++ env_path: "$ENV_PATH
@@ -76,11 +40,10 @@ RUN chgrp -R webgroup /srv
 RUN chmod -R u=rwX,g=rwX,o=X /srv
 
 # install python requirements
-RUN pip install -r "${WDIR}/requirements.txt"
+RUN pip install -r "${SRC_DIR}/osf_scraper_api/requirements.txt"
 
 # Expose ports for nginx
 EXPOSE 80 443
 
-# start app (xvfg is necessary for selenium, and supervisor is used to manage all other processes)
+# start app
 CMD (/usr/bin/supervisord -c /etc/supervisor/conf.d/supervisord.conf)
-#CMD (service xvfb start; export DISPLAY=:10; cd /srv/app; python /srv/app/osf_scraper_api/test.py)
