+++ osf_scraper_api/docker-compose.yml
@@ -1,9 +1,15 @@
   version: '3'
   services:
+    osf_scraper_base:
+      image: 'osf_scraper_base'
+      build:
+        context: ../
+        dockerfile: osf_scraper_api/BaseDockerFile
     osf_scraper_api:
       image: '${BUILD_ENV}_osf_scraper_api'
       build:
-        context: .
+        context: ../
+        dockerfile: osf_scraper_api/Dockerfile
         args:
           - BUILD_ENV
       ports:
@@ -12,12 +18,14 @@
       links:
         - redis
         - selenium
+      depends_on:
+        - osf_scraper_base
     redis:
       image: redis
       ports:
         - "6379"
     selenium:
-      image: selenium/standalone-firefox
+      image: selenium/standalone-chrome
       ports:
         - "4444:4444"
