+++ osf_scraper_api/osf_scraper_api/requirements.txt
@@ -6,8 +6,8 @@ rq==0.8.0
 redis==2.10.5
 rq-dashboard==0.3.8
 rq-scheduler==0.7.0
-git+git://github.com/mhfowler/open-source-feeds.git#egg=osf&subdirectory=osf
 uwsgi==2.0.15
 virtualenv==15.1.0
 flask-mail==0.9.1
 boto3==1.4.7
+selenium==3.5.0
