+++ osf_scraper_api/prod-docker-compose.yml
@@ -13,6 +13,6 @@
       ports:
         - "6379"
     selenium:
-      image: selenium/standalone-firefox
+      image: selenium/standalone-chrome
       ports:
         - "4444:4444"
